+++ crates/miden-objects/src/account/storage/map/mod.rs
@@ -11,6 +11,9 @@ use crate::{
     errors::StorageMapError,
 };
 
+mod partial;
+pub use partial::PartialStorageMap;
+
 // ACCOUNT STORAGE MAP
 // ================================================================================================