+++ crates/miden-objects/src/account/storage/partial.rs
@@ -1,10 +1,13 @@
-use alloc::vec::Vec;
+use alloc::collections::{BTreeMap, BTreeSet};
 
-use miden_crypto::merkle::{InnerNodeInfo, SmtLeaf, SmtProof};
+use miden_crypto::{
+    Word,
+    merkle::{InnerNodeInfo, SmtLeaf},
+};
 use vm_core::utils::{Deserializable, Serializable};
 
 use super::{AccountStorage, AccountStorageHeader, StorageSlot};
-use crate::{AccountError, Word};
+use crate::{AccountError, account::PartialStorageMap};
 
 /// A partial representation of an account storage, containing only a subset of the storage data.
 ///
@@ -15,24 +18,35 @@ use crate::{AccountError, Word};
 pub struct PartialStorage {
     /// Commitment of the account's storage slots.
     commitment: Word,
-    /// Account's storage header, containing top-level slot values.
+    /// Account storage header.
     header: AccountStorageHeader,
-    /// Merkle proofs for a subset of the account's storage maps keys
-    storage_map_proofs: Vec<SmtProof>,
+    /// Storage partial storage maps indexed by their root, containing a subset of the elements
+    /// from the complete storage map.
+    maps: BTreeMap<Word, PartialStorageMap>,
 }
 
 impl PartialStorage {
-    /// Returns a new instance of partial storage with the specified header and storage map proofs.
+    /// Returns a new instance of partial storage with the specified header and storage map SMTs.
     ///
     /// The storage commitment is computed during instantiation based on the provided header.
-    pub fn new(header: AccountStorageHeader, storage_map_proofs: Vec<SmtProof>) -> Self {
-        let commitment = header.compute_commitment();
-        PartialStorage { header, storage_map_proofs, commitment }
-    }
+    /// Additionally, this function validates that the passed SMTs correspond to one of the map
+    /// roots in the storage header.
+    pub fn new(
+        storage_header: AccountStorageHeader,
+        storage_maps: impl Iterator<Item = PartialStorageMap>,
+    ) -> Result<Self, AccountError> {
+        let storage_map_roots: BTreeSet<_> = storage_header.map_slot_roots().collect();
+        let mut maps = BTreeMap::new();
+        for smt in storage_maps {
+            // Check that the passed storage map partial SMT has a matching map slot root
+            if storage_map_roots.contains(&smt.root()) {
+                return Err(AccountError::StorageMapRootNotFound(smt.root()));
+            }
+            maps.insert(smt.root(), smt);
+        }
 
-    /// Returns a reference to the storage map proofs of this partial storage.
-    pub fn storage_map_proofs(&self) -> &[SmtProof] {
-        &self.storage_map_proofs
+        let commitment = storage_header.compute_commitment();
+        Ok(Self { commitment, header: storage_header, maps })
     }
 
     /// Returns a reference to the header of this partial storage.
@@ -45,32 +59,25 @@ impl PartialStorage {
         self.commitment
     }
 
-    /// Returns the value of the storage slot at the specified slot index.
-    ///
-    /// # Errors:
-    /// - If the index is out of bounds
-    pub fn get_item(&self, index: u8) -> Result<Word, AccountError> {
-        self.header.slot(index as usize).map(|(_type, value)| *value)
-    }
-
     // TODO: Add from account storage with (slot/[key])?
+
+    // ITERATORS
+    // --------------------------------------------------------------------------------------------
 
     /// Returns an iterator over inner nodes of all storage map proofs contained in this
     /// partial storage.
     pub fn inner_nodes(&self) -> impl Iterator<Item = InnerNodeInfo> {
-        // SAFETY: any u64 value is a valid SMT leaf index
-        self.storage_map_proofs.iter().flat_map(|proof| {
-            proof
-                .path()
-                .authenticated_nodes(proof.leaf().index().value(), proof.leaf().hash())
-                .expect("invalid SMT leaf index")
-        })
+        self.maps.iter().flat_map(|(_, map)| map.inner_nodes())
     }
 
-    /// Returns an iterator over leaves of all storage map entries contained in this partial
-    /// storage.
-    pub fn leaves(&self) -> impl Iterator<Item = &SmtLeaf> {
-        self.storage_map_proofs.iter().map(SmtProof::leaf)
+    /// Iterator over every [`PartialStorageMap`] in this partial storage.
+    pub fn maps(&self) -> impl Iterator<Item = &PartialStorageMap> + '_ {
+        self.maps.values()
+    }
+
+    /// Iterator over all tracked, non‑empty leaves across every map.
+    pub fn leaves(&self) -> impl Iterator<Item = &SmtLeaf> + '_ {
+        self.maps().flat_map(|map| map.leaves()).map(|(_, leaf)| leaf)
     }
 }
 
@@ -80,24 +87,24 @@ impl From<&AccountStorage> for PartialStorage {
     /// This creates a partial storage that contains proofs for all key-value pairs
     /// in all map slots of the account storage.
     fn from(account_storage: &AccountStorage) -> Self {
-        let mut storage_map_proofs = Vec::with_capacity(account_storage.slots().len());
+        let mut map_smts = BTreeMap::new();
         for slot in account_storage.slots() {
             if let StorageSlot::Map(map) = slot {
-                let proofs: Vec<SmtProof> = map.entries().map(|(key, _)| map.open(key)).collect();
-                storage_map_proofs.extend(proofs);
+                let smt: PartialStorageMap = map.clone().into();
+                map_smts.insert(smt.root(), smt);
             }
         }
 
         let header: AccountStorageHeader = account_storage.to_header();
         let commitment = header.compute_commitment();
-        PartialStorage { header, storage_map_proofs, commitment }
+        PartialStorage { header, maps: map_smts, commitment }
     }
 }
 
 impl Serializable for PartialStorage {
     fn write_into<W: vm_core::utils::ByteWriter>(&self, target: &mut W) {
         target.write(&self.header);
-        target.write(&self.storage_map_proofs);
+        target.write(&self.maps);
     }
 }
 
@@ -106,10 +113,10 @@ impl Deserializable for PartialStorage {
         source: &mut R,
     ) -> Result<Self, vm_processor::DeserializationError> {
         let header: AccountStorageHeader = source.read()?;
-        let storage_map_proofs: Vec<SmtProof> = source.read()?;
+        let map_smts: BTreeMap<Word, PartialStorageMap> = source.read()?;
 
         let commitment = header.compute_commitment();
 
-        Ok(PartialStorage { header, storage_map_proofs, commitment })
+        Ok(PartialStorage { header, maps: map_smts, commitment })
     }
 }